backend/requirements.txt
fastapi 
uvicorn 
pymongo 
python-jose 
passlib[argon2]
pydantic[email]
